moving the minting file
